[lib] allow missing `messageIDs` while comparing SQLite with `redux-persist`
Summary:
This allows `threadID` that is present in database but not redux. This will help while debugging - instead of error that `cannot read messageIDs of undefined` we will have more readable message, having information which keys are different. Is it issue for one key and e.g. it was not removed from DB while needed or there is some issue and `redux-persist` state was purged.

Depends on D7611

Test Plan:
Make inconsistency and check if it is properly detected and logged

Reviewers: tomek, marcin, michal