[keyserver] Introduce responder validators
Summary:
Introduce validators for entry, keys, siwe, message report, relationship and search responders and tests, most of which are taken from real requests that I've logged. I decided to not split it in multiple diffs because it's mostly just 1-2 validators per responder type and each change is localized to the responder.

Test Plan:
`yarn jest`

Reviewers: kamil, tomek

Subscribers: