[keyserver] Implement endpoint to upload external blob metadata
Summary:
For Blob service hosted multimedia, we don't store their content in the uploads table, but still need to store their metadata (dimensions, encryption key etc). This diff implements the endpoint code to upload this metadata. In contrast to the `upload_multimedia`, both input and output are JSON so we can use tcomb instead of manual validation.

The endpoint will be exposed in the next diff.

Depends on D7618

Test Plan: This is tested in the next diff in the stack.

Reviewers: atul, marcin, inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7619