[native] Display the number of pinned messages in a banner across chat
Summary: We want to display the number of pinned messages in a banner on top of chat the displays the number of pinned messages (if > 0 and if threadInfo contains a pinnedCount field since we gate it by
`hasMinCodeVersion` on native).

The code is largely similar to on `web`: https://phab.comm.dev/D7310

Depends on D7608

Test Plan: Check the banner to confirm the banner does not show if there are no pinned messages, and confirm that the banner appears for both 1 and more than 1 pinned messages.

Reviewers: atul, ginsu, tomek

Subscribers: