[lib] make fields in setting client store action optional
Summary:
I want to reuse this logic on the web, but only drafts are stored in the web database for now. Dispatching this with empty objects will affect the store, so we have somehow tell reducer that values are missing.

A different approach will be introducing a new action type, with only drafts, and using it in draft reducer only - in reviewers think that this is better I am open to updating this.

Depends on D7662

Test Plan:
Check on native is this does not affect current logic

Reviewers: tomek, marcin, michal