[native] move selectFromGalleryAndUpdateUserAvatar from avatar utils to edit user avatar provider
Summary: Introduce the edit user avatar provider. Another provider for the thread avatars will be introduced in a subsequent diff.

Test Plan: I tested all the avatar options we had (ENS, Emoji, and Photo gallery) and in all cases the behvaoir was to be expected with setting the avatar, loading while the avatar was saved

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7621