[native] introduce user avatar camera modal
Summary: Introduce user avatar camera modal

Test Plan:
Difficult to provide a demo video, but I was able to use my camera to take a photo and the photo i snapped was set as my avatar after hitting "send" in the camera modal. Also tested going back, and when I hit the "X" button it just goes back to the profile screen. I also double checked that the loading state was in sync with the edit user avatar
provider

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7624