[native] introduce edit thread avatar provider
Summary: In this diff I introduce the edit thread avatar provider and moved the remove user avatar functionality from the avatar-hooks to this provider. Future diffs will bring the other edit avatar functionality over to this provider

Test Plan:
Able to remove an avatar and the behavior is to be expected with removing an avatar and the loading state returns the correct values when the set avatar request was happening. Also tested that the
loaders were not being shared across different threads

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7629