[native] hardcode avatar feature flag to true
Summary: This is a naive/quick approach to hardcoding the avatar feature flag to true. Ideally we should not be using a hook anymore, but making that change would require me to go to every place this feature flag is used and update it accordingly. I personally felt that was a bit unnecessary especially after we talked today about deprecating all the avatar feature flags in the codebase sometime soon and this would take care of this using a unnecessary hook concern.  ([[ https://linear.app/comm/issue/ENG-3413/factor-out-copy-pasted-feature-flag-logic-code | ENG-3413 ]] considers removing and deprecating the feature flag rendering logic throughout our codebase)

Test Plan: Avatars now always renders on native

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7641