[native] hardcode avatar feature flag to true
Summary: This is a naive/quick approach to hardcoding the avatar feature flag to true. Ideally we should not be using a hook anymore, but making that change would require me to go to every place this feature flag is used and update it accordingly.

Test Plan: Avatars now always renders on native

Reviewers:

Subscribers: