[keyserver, lib] Add message search endpoint
Summary:
issue: https://linear.app/comm/issue/ENG-3170/create-a-keyserver-endpoint-for-the-message-search-within-specific
This endpoint allows searching messages. For now we can only search in one thread. It will have to be changed a bit in the future to be able to search in many threads, but  we don't yet know how we will want that to work.

Test Plan:
Tested with the next diff

Reviewers:
kamil, bartek, kuba
Subscribers: