[native] introduce edit user avatar provider scaffolding
Summary:
First step in breaking down D7621. In this diff I introduce the Edit User Avatar provider and moved the remove user avatar functionality from the avatar-hooks to this provider. Future diffs will bring the other edit avatar functionality over to this provider

Depends on D7573

Test Plan: Able to remove an avatar and the behavior is to be expected with removing an avatar and the loading state returns the correct values when the set avatar request was happening

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7626