[native] Display the pinned messages for the thread in the new screen
Summary: This is the screen that should display the pinned messages / the search results.
The styling will either come in a revision to this diff or in the next diff, but I'm
putting this up for now to unblock message searching.

Note: This diff is a rough implementation of this screen, it's not had a chance to be
refined yet.

Depends on D7672

Test Plan: Check to see the pinned messages render in the screen, despite the styling
being incorrect for them.

Reviewers: atul, ginsu, tomek

Subscribers: inka