[native] have saving thread avatar loaders only appear for thread being updated
Summary: Added the `activeThreadID`/`threadID` to the thread settings loading status selector custom keys. This fixes the issue where if we update a thread avatar the loaders would be shared and shown across different threads.

Test Plan: Tested that the loader only appeared in the thread avatar I was updating:

Reviewers:

Subscribers: