[keyserver] Mark ids in input validators
Summary:
We will also need to convert ids in the input, so this diff marks the ids for conversion inside existing input validators. `tID` should behave the same as `t.String` so there should be no runtime difference.

Depends on D7712

Test Plan: Test if the web and mobile app still work.

Reviewers: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7713