[keyserver] Convert responders output
Summary:
Use the validators and the `validateOutput` function introduced in the previous diff in the api responders. The conversion is behind the flag for now, but the outputs are still validated.

Depends on D7711

Test Plan:
Test web and ios emulator if all outputs are validated correctly and no server error is thrown.

Reviewers: tomek, kamil

Subscribers: