[native] introduce edit thread avatar provider
Summary:
In this diff I introduce the edit thread avatar provider and moved the remove user avatar functionality from the avatar-hooks to this provider. Future diffs will bring the other edit avatar functionality over to this provider.

Depends on D7624

Test Plan: Able to remove an avatar and the behavior is to be expected.

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7629