[web] Upload encrypted media to Blob service
Summary:
Web counterpart of D7650. This diff that enables encrypted media uploads to Blob service on web.
I decided to hide this feature behind a boolean flag to be able to easily flip the switch.

Depends on D7688, D7689

Test Plan:
Set the useBlobServiceUploads to true. Send a message with single and multiple images. Verify they're sent successfully. Verify Blob service content (S3 + DynamoDB) and keyserver uploads table.
When the pending upload is deleted (user clicks the cross in the corner of selected img), the blob is removed from blob servic of selected img), the blob is removed from blob service

Reviewers: atul, marcin

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7690