[landing] update header strucutre to use flexbox over grid
Summary: To update the `Header` component to match the redesigns, I broke it up into three diffs. This is the first diff. This diff removes the grid layout in favor for a flexbox layout. I chose to do this

Test Plan: The header looks and behaves as expected when resizing the window to different widths

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7835