[native] Fix ComposedMessage width constraint to only include content
Summary:
The avatar should not be part of the width constraint. The width constraint is only meant to apply to the content. This change fixes [ENG-3826](https://linear.app/comm/issue/ENG-3826/message-with-bullet-points-overflows-its-bounds-on-native), and makes the width match for non-viewer and viewer (no avatar) messages.

Our ability to use this fix is actually kind of a coincidence... we luckily don't need to include `avatar` here, but if we did it appears that there might be a React Native bug. Including this simplified non-working example for future reference: https://gist.github.com/Ashoat/f0f00496d17db8df18c2cf21f753f469

Depends on D7697

Test Plan: [screenshot]

Reviewers: ginsu, atul

Subscribers: