[native] Unlink files as necessary in `useUploadSelectedMedia`
Summary:
`expo-image-picker` will always copy the contents of image into cache and provide a URL to cached image. As a result we'll ALWAYS need to unlink the `selection.uri`.

We also clean up processedMedia.uploadURI if necessary.

Test Plan: Tested small images which weren't processed and large images which were processed. Logged the contents of `selection`, `processedMedia`, and `urisToBeDisposed`. When I commented out the `unlink` line (159), I was able to open all of the images in Finder. When I uncommented the `unlink` line, all of the images were inaccessible because they'd been removed:

Reviewers: ashoat, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7704