[landing] introduce new typography design system
Summary: Introduce a more designer friendly typography design system. The goal of this is to create a very clear source of truth between ted/future designers and any dev when it comes to which typogprahy we use in our designs and code.  This file was built by using Ted's new typography design system in the figma

Test Plan: Had Ted review this, and got his stamp of approval, and I also locally tested every typography and made sure that they matched what was in the designs

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7755