[keyserver] Add migration to populate `target_message` for sidebar_source
Summary:
issue: https://linear.app/comm/issue/ENG-3708/add-a-migration-to-populate-target-message-for-sidebar-source
Adding a migration that sets the `target_message` column for sidebar_source messages to the id of the original message, that we store in the content JSON as well.

Test Plan: Run keyserver, checked that `taget_message` got set for messages with type 17.

Reviewers: kamil, bartek, ashoat

Reviewed By: bartek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7517