[native] Don't crash if TagInput is unmounted before error received
Summary:
I saw a bug where if I added a user, then left the screen before the server response came in, BUT the server response was an error, I would get a crash with "tagInput should be set".

The issue is that we have an `invariant` that assumes the `TagInput` is mounted when the error comes in. But if the user leaves the screen before the error, then the `TagInput` might be unmounted before the error.

Test Plan: Confirm there are no remaining instances of `tagInput should be set` in the codebase

Reviewers: atul, ginsu, tomek

Subscribers: