[web] Introduce barebones `CommunityCreationModal`
Summary:
Absolutely minimal `CommunityCreationModal` that opens when the `CommunityCreationButton` is clicked and is dismissed when either A. the X is clicked B. the user clicks outside of the Modal.

In subsequent diffs each part of the modal will be added piecewise.

Test Plan:
Looks and behaves as expected.
https://blob.sh/f417ec.png

Reviewers:
ashoat, ginsu