[landing] introduce colors from design system
Summary:
This diff introduces all the colors in Ted's new color design system. For this diff I decided to just add the colors and leave the other variables in for now just so I don't cause any regressions I may be unaware about. As I keep working on the landing page, I will slowly phase out and get rid of any unused colors/variables. One thing to note is that this diff does not include `black-95`. I was instructed by Ted not to include this since this color is only used in the nav sidebar and Ted said that he just put it in here to record it, but will eventually figure out a better place for that color

Links to color design system:
https://www.figma.com/file/fwKJ6eNgfxlx8a7WxSyAZL/2023-Design-System?node-id=1-427&t=ds82zlYncjLEbFB9-4

Link to landing page specific colors:
https://www.figma.com/file/HAAQwcEIjM32gnEMy2d0k5/Comm-%2F-Landing?node-id=3263-41925&t=EVfPbbD9hMG9rkrB-4

Depends on D7755

Test Plan: Had ted double check this, and got his stamp of approval

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ted, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7756