[native] Display a button and navigate to invite links if they are present
Summary: When a community has a primary link, display three dots button next to it. The links are updated on every drawer open. Clicking this link shows a button that can be used to open a modal with the links.

Depends on D7730

Test Plan: Create a primary link for one community. Check if the button is displayed next to it. Click the button and open invite links modal - it should contain the link.

Reviewers: bartek, kamil, inka