[native] Create a new screen to display the thread's pinned messages
Summary:
Create a very basic new screen to show the thraed's pinned messages (this screen
will be re-used by the message search work on native).

Implementation of the screen will come in the next diff, to make this one easier to
review.

Depends on D7614

Test Plan:
Navigate to the screen by clicking on the pinned count banner from the previous
diff and make sure the screen is navigated to.

Reviewers: atul, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7672