[lib] fix `callServerEndpoint` returning `undefined`
Summary:
This is fix for some serious issue - context in [ENG-3806](https://linear.app/comm/issue/ENG-3806/callserverendpoint-can-return-undefined).
When policies were not accepted this function was returning `undefined` instead of throwing an error to inform the caller that the request did not succeed.

Test Plan:
Test policy acknowledgment flow on web, native, and desktop in multiple scenarios.

Reviewers: tomek, michal