[native] Show a pin icon next to pinned messages
Summary:
Next to each pinned message, we want to display a pin icon to indicate that the message is pinned. We only want to show this icon in chat, and not the pin modal / pinned messages screen.

Logic is pretty much identical to https://phab.comm.dev/D7505

Depends on D7735

Linear: https://linear.app/comm/issue/ENG-3455/show-a-pin-icon-next-to-pinned-messages

Test Plan:
Test to confirm that the pin indicator appears in chat and not the other instances where `Message` is rendered (like the pin modal and screen).

For iOS, the confirmation is that there are no message height inconsistencies in the terminal. For Android, they appear on `master`, but the message height warnings don't show for all of the pinned messages
so it makes me believe no changes made in this diff cause the warnings.

Reviewers: atul, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7742