[native] Check install referrer on Android
Summary:
When a user clicks an invite link and doesn't have the app installed, we redirect him to Play Store from where the app can be installed. This redirection specifies a referrer parameter which can be used to check from where the app was installed. The, when the app is run, we check if referrer was set and if that's the case, we use it just like an invite link.

Depends on D7751

Test Plan: Open an invite link on an emulator without the app installed. Install the app. Run the app - the invitation modal should appear. Close the app and open it - the invitation shouldn't appear. Log out, close the app, open the app, log in - the invitation shouldn't appear.

Reviewers: bartek, kamil, inka, ashoat

Reviewed By: bartek, ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7744