[lib][native][web] Fix bug where no messages are loaded if most recent 20 are reactions
Summary:
This resolves [ENG-3851](https://linear.app/comm/issue/ENG-3851/when-all-most-recent-messages-are-reactions-fetching-older-messages). There are two issues here:

1. `oldestMessageServerID` is checking `messageListData`, which does not include reactions by themselves. If new results are all reactions, `messageListData` will not update, and `oldestMessageServerID` will stay the same.
2. The code for setting `loadingFromScroll` relies on `messageListData` changing, which (as mentioned above) is no longer guaranteed if returned messages are reactions.

Test Plan: I tested this on both web and native by creating over 20 reactions in my local environment. That allowed me to repro the issue after an initial load on web or a log-in on native. I was able to confirm the issue was resolved with this diff

Reviewers: atul, tomek, ginsu

Differential Revision: https://phab.comm.dev/D7746