[native] Show message tooltip on press in the pinned messages screen
Summary:
We need to measure vertical bounds in order to display the tooltip within the pinned messages screen. Confirming functionality of each tooltip action will come in the following diff.

Adding @ashoat since this is the approach we discussed in our sync.

Linear: https://linear.app/comm/issue/ENG-3460/show-message-tooltip-on-press-in-the-pinned-messages-screen

Depends on D7673

Test Plan: Please watch the video to show the tooltip now appearing in the screen

Reviewers: ashoat, atul, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7735