[web] Always disable default authorName and always show avatars in message results modal
Summary:
This isn't urgent and not blocking anything, just something I noticed when viewing
pinned messages in the modal on web. Sometimes, the default authorName
is shown (we want to avoid that since we render one more lined up with the
Figma designs and its rendering is conditional so it won't always appear),
and sometimes the avatars don't show up.

Two really small changes here, both are already done on native:

- In the modified object, sets startsCluster: true for the username
- In the modified object, sets endsCluster: true for the avatar

Test Plan:
Confirm the before and avatar in the modal to validate the changes

Before:
{F528021}

After:
{F528020}

Reviewers: atul, ginsu, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7747