[native] Introduce a hook that allows running an effect just once per installation
Summary:
When checking referrer on Android, we would like to do it at most once per installation. To achieve that we can set a value in `AsyncStorage` and run the effect only if the value is not set.

Depends on D7743

Test Plan: Tested in combination with the next diff. Install the app using the referrer on Android, open the app and cancel invitation, close and open the app - the invitation shouldn't appear again.

Reviewers: bartek, kamil, inka, ashoat

Reviewed By: bartek, ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7751