[landing] introduce mobile nav compoment
Summary:
This diff introduces the mobile nav component and the styles to go with it. This diff is just the component and the next couple of diffs will introduce the animation for it and integration of it into the landing page.

[[ https://www.figma.com/file/HAAQwcEIjM32gnEMy2d0k5/Comm-%2F-Landing?node-id=3268-47759&t=EVfPbbD9hMG9rkrB-4 | Figma ]]

Linear Task: https://linear.app/comm/issue/ENG-3424/update-the-nav-header

Depends on D7759

Test Plan:
Here are some screenshots to show what the mobile nav component will look like when it is active (the red border is only in the screenshot just to make it easier for reviewers to see what this component is)

{F530642}

{F530641}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7763