[native] Reformat the file so that it's easier to update it
Summary:
In the next diff I have to update this file with a new permission. At this point it is not feasible, as I would need to perform 150 changes. One easy option to update this file could be to console log the expected content and then format it using Eslint. But in order to do that the file has to be formatted using the same strategy from the beginning.

This diff modifies this file by changing how it is formatted. I console logged the content, pasted it and formatted using Eslint. The same strategy can be used in the future to update it.

Test Plan: Run the tests.

Reviewers: bartek, kamil, inka, atul