[keyserver] Create an endpoint that can be used to create or update a public link
Summary: If a community doesn't have a public invite link - it will be created. If it has one, it will be updated.

Depends on D7762

Test Plan: Create a new community, create a link for it using manage public link screen (later in the stack) - a link should be created. Update it, using the same screen - it should be updated.

Reviewers: bartek, kamil, inka