[lib] Add client side boilerplate and reducer for an endpoint that allows creating or updating public links
Summary:
The only interesting part is that the link in redux is updated after successful call to this endpoint. For that purpose I'm using a link returned by backend, instead of figuring out the values on the client, because at some point we could consider having different defaults.

Depends on D7878

Test Plan: Tested in combination with the rest of the stack. Checking if updating or creating a public link works is enough to verify the correctness.

Reviewers: bartek, kamil, inka

Reviewed By: kamil

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7879