[landing] animate mobile nav component
Summary:
This diff animates the mobile nav component from the top to the bottom. I worked closely with Ted on this to get the animation exactly how he wanted it to look/feel. Some things to note is that this diff is safe to land in its current state. The mobile nav component is currently hidden above the screen where it lives until someone presses the hamburger icon which will be introduced in subsequent diffs. Furthermore the first demo in the test plan may look a bit funky and that is because we need to make changes to the `Header` component as well. I will attach a second demo which will preview the "final form".

**The focus of this diff is purely the up and down animation the mobile nav bar does**

Depends on D7763

Test Plan:
Please watch the demo videos below to see how the up and down animation of the mobile nav bar looks and feels

{F533184}

Sneak Peak of the animation with a more polished header:

{F533185}

Reviewers: atul, kamil, ashoat

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7781