[native] Use MessageSearchContent
Summary:
issue: https://linear.app/comm/issue/ENG-3163/create-the-search-screen-for-message-searching

Test Plan:
Checked that the messages are fetched and displayed correctly. Checked that the "End of results" and "No results..." messages display correctly. Checked the case when
the first fetch returns all results as well as the case when the user needs to scroll to fetch more messages a couple of times. Checked that if the query is cleared "Your search results will appear here!" message is displayed.
Checked that if the query is not cleared between the searches everything works - only the results of the second query are displayed.

Reviewers:
kuba, kamil, bartek

Subscribers: