[native] Add thumbhash step to media mission
Summary:
This uses the expo-module to generate the thumbhash for the media file during the media mission. For videos, the thumbnail is used.
For encrypted media, the thumbhash field is also encrypted using the photo/thumbnail encryption key.

Depends on D7780

Test Plan: This will be tested more with subsequent diffs. Now console.logged the thumbhash field and sent some media.

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7802