[web] Display edit state (in the background)
Summary:
Display the edit message component.

Later I will add a modal, which covers the screen, to prevent the user from navigating outside the thread. This one is to move other messages in chats accordingly to the height of this component.

Test Plan:
Run the app. Press the edit button. It shows the new component. Pressed cancel in the new component. It left edit state.

Screenshot:
{F540367}

Reviewers: michal, inka

Reviewed By: michal

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7827