[lib] add local id field to existing reports
Summary:
This is probably not necessary as I don't think that anyone has some persisted, not sent reports, but adding this for clarity.

Depends on D7910

Test Plan:
Make sure that there are some reports in store (e.g. throw in keyserver endpoint and create some inconsistencies). Then verify that `id` was added.

Reviewers: tomek, michal